[MA-0] MovieGenresUserDefaults class and UserDefaultsConstants moved related class and structs into separate file
