+++ movieApp/movieApp/Utility/UserDefaultsManager/UserDefaultsConstants.swift
@@ -1,0 +1,12 @@
+//
+//  UserDefaultsConstants.swift
+//  movieApp
+//
+//  Created by Can Kurtur on 8.09.2021.
+//
+
+import Foundation
+
+struct UserDefaultsConstants {
+    static let genresDict = "genresDict"
+}
